Setup typescript environment
The language for this excercise will be TypeScript. To use it to its
fullest potential, we'll be setting up the TSCompiler configuration and
TSLint rules. This should assure us we're not making some simple
mistakes and that the compiler will tell us if something would not be
working as expected.

The `package.json` defines few simple tasks, to make the developer's
lives easier, such as `build`, `start`, `watch`.