Implement Sierpinski logic
We're delivering a small class, that will help us in implementation of
Sierpinski triangle on the fly. It uses the `d3` library previously
installed with NPM. The class, makes a single assumption, which is the
amount of levels deep, the triangles will be drawn. At the current
state, this is not managable from the API interface.

We're installing the d3 library, which is a nice helper making the
management of HTML, CSS and SVG easier for our purpose - which is to
generate the Sierpinski triangle.