Compile the application
Now that the logic code is ready, we can compile the code into JS that
will be understood by our browser. However in order to achieve that, our
HTML file will need to load the SystemJS library that will be able to
manage the modules in the way we'd like them to work. This is a drawback
we have to take care of, due to the initial laziness and not
implementing more tooling around.

We're installing systemjs library, as our TypeScript is compiled for
that environment. It will asure our code can be run in the web browser
when loaded.